requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.0
nltk>=3.6.0
python-dotenv>=0.19.0
openai>=0.10.2
ollama
pytest>=6.2.5
